Backend/qaba-api/pyproject.toml
[project]
name = "qaba-api"
version = "1.0.0"
description = "QABA Real Estate Platform Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Django core
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",

    # Django extensions
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-otp>=1.6.1",
    "django-two-factor-auth>=1.18.1",
    "django-formtools>=2.5.1",
    "django-phonenumber-field>=8.1.0",
    "django-extra-fields>=3.0.2",

    # DRF extensions
    "drf-nested-routers>=0.94.1",
    "drf-spectacular>=0.28.0",

    # Storage & Media
    "cloudinary>=1.42.2",
    "django-cloudinary-storage>=0.3.0",
    "pillow>=11.0.0",

    # Database
    "psycopg2-binary>=2.9.10",

    # Server & Static files
    "gunicorn>=23.0.0",
    "whitenoise>=6.9.0",

    # Authentication & Security
    "google-auth>=2.41.1",
    "msal>=1.33.0",
    "cryptography>=46.0.1",
    "pyjwt>=2.10.1",

    # Utilities
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "qrcode>=8.2",
    "phonenumbers>=8.13.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",

    # Code quality
    "ruff>=0.8.0",
    "mypy>=1.13.0",

    # Django dev tools
    "django-debug-toolbar>=4.4.0",
    "django-extensions>=3.2.3",

    # Type stubs
    "django-stubs>=5.1.0",
    "djangorestframework-stubs>=3.15.0",
    "types-requests>=2.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "core"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "django-debug-toolbar>=4.4.0",
    "django-extensions>=3.2.3",
    "django-stubs>=5.1.0",
    "djangorestframework-stubs>=3.15.0",
    "types-requests>=2.32.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    ".venv",
    "venv",
    "staticfiles",
    "media",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "core"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings.development"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
ignore_missing_imports = true
strict = false
warn_unused_ignores = true
exclude = ["migrations", "staticfiles", "media"]

[tool.django-stubs]
django_settings_module = "core.settings.development"
